examples/rag_demo/requirements.txt
qdrant-client
httpx
apple-fm-cli
pypdf
fpdf2
langchain-text-splitters
